# services/backend-gateway/docs/static/openapi/quckchat-api.yaml
openapi: 3.0.3
info:
  title: QuckChat API
  description: |
    Real-time messaging platform API with audio/video calling capabilities.

    ## Features
    - Phone OTP authentication
    - JWT-based authorization
    - Real-time messaging via WebSocket
    - Voice and video calls with WebRTC
    - File uploads (S3/local storage)
    - Push notifications (Firebase FCM)

    ## Authentication
    Most endpoints require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```

    ## Rate Limiting
    - Default: 100 requests per minute
    - Auth endpoints: 10 requests per minute
    - OTP: 3 requests per minute

    ## WebSocket Endpoints
    - Chat: `ws://localhost:3000/chat`
    - WebRTC: `ws://localhost:3000/webrtc`
  version: 1.0.0
  contact:
    name: QuckChat Support
    email: support@quckchat.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.quckchat.com/api/v1
    description: Production server
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Get a new access token using a refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout user
      description: Invalidate the current session
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out successfully
...
  /auth/sessions:
    get:
      tags: [Auth]
      summary: Get active sessions
      description: List all active login sessions
      operationId: getSessions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
  /auth/sessions/{sessionId}:
    delete:
      tags: [Auth]
      summary: Terminate session
      description: Terminate a specific session
      operationId: terminateSession
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session terminated

  # ==================== USERS ====================
  /users/me:
    get:
      tags: [Users]
      summary: Get current user
      description: Get the authenticated user's profile
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags: [Users]
      summary: Update profile
      description: Update the authenticated user's profile
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/search:
    get:
      tags: [Users]
      summary: Search users
      description: Search users by name or phone number
      operationId: searchUsers
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSearchResponse'
  /users/{userId}:
    get:
      tags: [Users]
      summary: Get user by ID
      description: Get a user's public profile
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== CONVERSATIONS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
        timestamp:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        _id:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatar:
          type: string
        bio:
          type: string
        status:
          type: string
          enum: [online, offline, away, busy]
        lastSeen:
          type: string
          format: date-time
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UpdateUserDto:
      type: object
      properties:
        displayName:
          type: string
          maxLength: 100
        username:
          type: string
          pattern: '^[a-zA-Z0-9_]+$'
        bio:
          type: string
          maxLength: 500
        avatar:
          type: string
    UserSearchResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Session:
      type: object
      properties:
        _id:
          type: string
        deviceInfo:
          type: object
          properties:
            deviceName:
              type: string
            platform:
              type: string
            ip:
              type: string
        isActive:
          type: boolean
        lastActive:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        totalDocs:
          type: integer
        hasNextPage:
          type: boolean
        hasPrevPage:
          type: boolean